+++ src/Model/PowerFlow/ModelEvaluatorImpl.hpp
@@ -30,18 +30,18 @@ namespace GridKit
       : size_(size),
         size_quad_(size_quad),
         size_opt_(size_opt),
-        y_(size_),
-        yp_(size_),
-        f_(size_),
-        g_(size_quad_),
-        yB_(size_),
-        ypB_(size_),
-        fB_(size_),
-        gB_(size_opt_),
+        y_(static_cast<size_t>(size_)),
+        yp_(static_cast<size_t>(size_)),
+        f_(static_cast<size_t>(size_)),
+        g_(static_cast<size_t>(size_quad_)),
+        yB_(static_cast<size_t>(size_)),
+        ypB_(static_cast<size_t>(size_)),
+        fB_(static_cast<size_t>(size_)),
+        gB_(static_cast<size_t>(size_opt_)),
         jac_(COO_Matrix<ScalarT, IdxT>()),
-        param_(size_opt_),
-        param_up_(size_opt_),
-        param_lo_(size_opt_)
+        param_(static_cast<size_t>(size_opt_)),
+        param_up_(static_cast<size_t>(size_opt_)),
+        param_lo_(static_cast<size_t>(size_opt_))
     {
     }
